requirements.txt
pandas==2.3.3
numpy==2.4.1
fastapi==0.128.0
uvicorn==0.40.0
streamlit==1.53.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pyarrow==23.0.0
sentence-transformers==5.2.1
torch==2.10.0
huggingface_hub==1.3.4
pydantic==2.12.5
hf_transfer==0.1.9
hf_xet==1.2.0
pytest==9.0.2
httpx==0.28.1
transformers==5.0.0
llama-cpp-python==0.3.2
langchain-huggingface==0.3.1
langchain==0.3.27
tqdm==4.67.1
